do not overwrite assets files if they exist
git-svn-id: https://rails-widgets.googlecode.com/svn/trunk@37 dbc01b36-f746-0410-8b88-c76c4e1ff24b

+++ install.rb
@@ -2,7 +2,7 @@ def copy(file_name, from_dir, to_dir)
   FileUtils.mkdir to_dir unless File.exist?(File.expand_path(to_dir))   
   from = File.expand_path(File.join(from_dir,file_name))
   to = File.expand_path(File.join(to_dir, file_name))
-  FileUtils.cp from, to, :verbose => true
+  FileUtils.cp from, to, :verbose => true unless File.exist?(to)
 end
 
 def copy_image(file_name)
